+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - Award: only the year is displayed.
   - Added Services.
   - Talks and events are now grouped per year, and sorted by date within each year.
+  - LaTeX: possibility to add class options.
 
 ## Bug fixes
   - Type annotations do not cause missing imports errors (#3).